perf(util): split lazyvim.util in smaller separate modules
