+++ lua/lazyvim/plugins/extras/editor/mini-files.lua
@@ -57,7 +57,7 @@ return {
     vim.api.nvim_create_autocmd("User", {
       pattern = "MiniFilesActionRename",
       callback = function(event)
-        require("lazyvim.util").on_rename(event.data.from, event.data.to)
+        require("lazyvim.util").lsp.on_rename(event.data.from, event.data.to)
       end,
     })
   end,
